Test e2e that sing-off enforcement fires
This is an end-to-end test to manually check that the sign-off
enforcements fires.

Please do not close this pull request. If you want to re-test the
sign-off enforcement, just rebase this commit to the master branch and
do a forced push.

No `Signed-off-by:` line is expected in this commit body and the
developer should manually check that the checks fail on this pull
request.