Discontinue local scripts (#98)
It is hard to add new features to the action and at the same time
maintain parallel implementation in PowerShell. In addition, local
scripts are usually not even used, so their utility is questionable in
the first place.

We decided to discontinue them in order to facilitate further
development in the future.